Added navbar, removed cancel button

+++ BloodBankSystem/storage/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, redirect
 from django.urls import reverse
 from django.views import View
 from django.contrib import messages
-from .forms import *
+from storage.forms import *
 
 
 # Create your views here.
@@ -10,14 +10,16 @@ class CreateBloodSupplyView(View):
     template = 'create_blood_supply.html'
 
     def get(self, request):
-        form = CreateBloodSupplyForm
-        return render(request, self.template, {'form': form})
+        if 'username' not in request.session:
+            return redirect(reverse('accounts:login'))
+        else:
+            form = CreateBloodSupplyForm
+            return render(request, self.template, {'form': form})
 
     def post(self, request):
         form = CreateBloodSupplyForm(request.POST)
         if form.is_valid():
             form.save()
-
             org = Organization.objects.get(username=request.session['username'])
             new_supply = BloodSupply.objects.latest('supply_id')
             org.blood_supply = new_supply
@@ -32,9 +34,12 @@ class UpdateBloodSupplyView(View):
     template = 'update_blood_supply.html'
 
     def get(self, request):
-        bloodSupply = BloodSupply.objects.get(pk=request.session['blood_supply_id'])
-        form = UpdateBloodSupplyForm(instance=bloodSupply)
-        return render(request, self.template, {'form': form})
+        if 'username' not in request.session:
+            return redirect(reverse('accounts:login'))
+        else:
+            bloodSupply = BloodSupply.objects.get(pk=request.session['blood_supply_id'])
+            form = UpdateBloodSupplyForm(instance=bloodSupply)
+            return render(request, self.template, {'form': form})
 
     def post(self, request):
         bloodSupply = BloodSupply.objects.get(pk=request.session['blood_supply_id'])
@@ -49,7 +54,11 @@ class UpdateBloodSupplyView(View):
 class DeleteBloodSupplyView(View):
 
     def get(self, request):
-        owned_supply = BloodSupply.objects.get(pk=request.session['blood_supply_id'])
-        owned_supply.delete()
-        request.session['blood_supply_id'] = None
-        return redirect(reverse('accounts:index'))
+        if 'username' not in request.session:
+            return redirect(reverse('accounts:login'))
+        else:
+            owned_supply = BloodSupply.objects.get(pk=request.session['blood_supply_id'])
+            owned_supply.delete()
+            request.session['blood_supply_id'] = None
+            messages.success(request, 'Blood supply deleted succesfully.')
+            return redirect(reverse('accounts:index'))
